+++ src/interfaces/libpq/fe-misc.c
@@ -25,7 +25,7 @@
  *
  *
  * IDENTIFICATION
- *	  $Header: /cvsroot/pgsql/src/interfaces/libpq/fe-misc.c,v 1.41 2000/04/12 17:17:15 momjian Exp $
+ *	  $Header: /cvsroot/pgsql/src/interfaces/libpq/fe-misc.c,v 1.42 2000/11/16 05:51:05 momjian Exp $
  *
  *-------------------------------------------------------------------------
  */
@@ -209,7 +209,7 @@ pqGetnchar(char *s, size_t len, PGconn *conn)
 	conn->inCursor += len;
 
 	if (conn->Pfdebug)
-		fprintf(conn->Pfdebug, "From backend (%d)> %.*s\n", len, (int) len, s);
+		fprintf(conn->Pfdebug, "From backend (%lu)> %.*s\n", (unsigned long)len, (int) len, s);
 
 	return 0;
 }
@@ -260,13 +260,13 @@ pqGetInt(int *result, size_t bytes, PGconn *conn)
 			break;
 		default:
 			sprintf(noticeBuf,
-					"pqGetInt: int size %d not supported\n", bytes);
+					"pqGetInt: int size %lu not supported\n", (unsigned long)bytes);
 			DONOTICE(conn, noticeBuf);
 			return EOF;
 	}
 
 	if (conn->Pfdebug)
-		fprintf(conn->Pfdebug, "From backend (#%d)> %d\n", bytes, *result);
+		fprintf(conn->Pfdebug, "From backend (#%lu)> %d\n", (unsigned long)bytes, *result);
 
 	return 0;
 }
@@ -297,13 +297,13 @@ pqPutInt(int value, size_t bytes, PGconn *conn)
 			break;
 		default:
 			sprintf(noticeBuf,
-					"pqPutInt: int size %d not supported\n", bytes);
+					"pqPutInt: int size %lu not supported\n", (unsigned long)bytes);
 			DONOTICE(conn, noticeBuf);
 			return EOF;
 	}
 
 	if (conn->Pfdebug)
-		fprintf(conn->Pfdebug, "To backend (%d#)> %d\n", bytes, value);
+		fprintf(conn->Pfdebug, "To backend (%lu#)> %d\n", (unsigned long)bytes, value);
 
 	return 0;
 }
